refactor: split db.rs into focused modules with colocated tests
Split the monolithic db.rs (1786 lines) into domain-focused submodules
with tests living next to the code they exercise:

- db.rs           — Database struct, open(), shared test helpers (168 lines)
- db/media.rs     — LibraryMedia trait impl, row types, tests (651 lines)
- db/albums.rs    — LibraryAlbums trait impl, tests (365 lines)
- db/sync.rs      — Sync upserts, checkpoints, audit (235 lines)
- db/thumbnails.rs — Thumbnail status queries (66 lines)
- db/stats.rs     — Library statistics (31 lines)
- db/upload.rs    — Upload queue queries (76 lines)

No API changes — all callers use Database methods identically.

Closes #164
