+++ src/library/db.rs
@@ -1,18 +1,19 @@
 use std::path::Path;
-
 use std::time::Duration;
 
 use sqlx::sqlite::{SqliteConnectOptions, SqliteJournalMode, SqlitePoolOptions};
 use sqlx::SqlitePool;
 use tracing::{info, instrument};
 
-use super::album::{Album, AlbumId, LibraryAlbums};
 use super::error::LibraryError;
-use super::media::{
-    LibraryMedia, MediaCursor, MediaFilter, MediaId, MediaItem, MediaMetadataRecord, MediaRecord,
-    MediaType,
-};
-use super::thumbnail::ThumbnailStatus;
+use super::media::{MediaId, MediaRecord};
+
+mod albums;
+mod media;
+mod stats;
+mod sync;
+mod thumbnails;
+mod upload;
 
 /// Aggregate library statistics for the preferences overview.
 #[derive(Debug, Clone, Default)]
@@ -72,990 +73,6 @@ impl Database {
     }
 }
 
-/// Internal row type for `media_metadata` queries.
-#[derive(sqlx::FromRow)]
-struct MetadataRow {
-    camera_make: Option<String>,
-    camera_model: Option<String>,
-    lens_model: Option<String>,
-    aperture: Option<f32>,
-    shutter_str: Option<String>,
-    iso: Option<i64>,
-    focal_length: Option<f32>,
-    gps_lat: Option<f64>,
-    gps_lon: Option<f64>,
-    gps_alt: Option<f64>,
-    color_space: Option<String>,
-}
-
-/// Internal row type for `list_media` — maps SQLite columns to Rust types.
-#[derive(sqlx::FromRow)]
-struct MediaRow {
-    id: String,
-    taken_at: Option<i64>,
-    imported_at: i64,
-    original_filename: String,
-    width: Option<i64>,
-    height: Option<i64>,
-    orientation: i64,
-    media_type: i64,
-    is_favorite: i64,
-    is_trashed: i64,
-    trashed_at: Option<i64>,
-    duration_ms: Option<i64>,
-}
-
-impl MediaRow {
-    fn into_item(self) -> MediaItem {
-        MediaItem {
-            id: MediaId::new(self.id),
-            taken_at: self.taken_at,
-            imported_at: self.imported_at,
-            original_filename: self.original_filename,
-            width: self.width,
-            height: self.height,
-            orientation: self.orientation as u8,
-            media_type: if self.media_type == 1 {
-                MediaType::Video
-            } else {
-                MediaType::Image
-            },
-            is_favorite: self.is_favorite != 0,
-            is_trashed: self.is_trashed != 0,
-            trashed_at: self.trashed_at,
-            duration_ms: self.duration_ms.map(|v| v as u64),
-        }
-    }
-}
-
-impl Database {
-    /// Insert a `Pending` thumbnail row. No-op if a row already exists.
-    pub async fn insert_thumbnail_pending(&self, id: &MediaId) -> Result<(), LibraryError> {
-        let id_str = id.as_str();
-        sqlx::query(
-            "INSERT OR IGNORE INTO thumbnails (media_id, status) VALUES (?, 0)",
-        )
-        .bind(id_str)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Mark a thumbnail `Ready` and record its relative `file_path`.
-    pub async fn set_thumbnail_ready(
-        &self,
-        id: &MediaId,
-        file_path: &str,
-        generated_at: i64,
-    ) -> Result<(), LibraryError> {
-        let id_str = id.as_str();
-        sqlx::query(
-            "UPDATE thumbnails SET status = 1, file_path = ?, generated_at = ? WHERE media_id = ?",
-        )
-        .bind(file_path)
-        .bind(generated_at)
-        .bind(id_str)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Mark a thumbnail `Failed`.
-    pub async fn set_thumbnail_failed(&self, id: &MediaId) -> Result<(), LibraryError> {
-        let id_str = id.as_str();
-        sqlx::query("UPDATE thumbnails SET status = 2 WHERE media_id = ?")
-            .bind(id_str)
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Return the `relative_path` column for `id`, or `None` if no row exists.
-    ///
-    /// Fetch a single media item by ID.
-    ///
-    /// Used for incremental grid updates without full reload.
-    pub async fn get_media_item(&self, id: &MediaId) -> Result<Option<MediaItem>, LibraryError> {
-        let row: Option<MediaRow> = sqlx::query_as(
-            "SELECT id, taken_at, imported_at, original_filename,
-                    width, height, orientation, media_type, is_favorite,
-                    is_trashed, trashed_at, duration_ms
-             FROM media WHERE id = ?",
-        )
-        .bind(id.as_str())
-        .fetch_optional(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(row.map(MediaRow::into_item))
-    }
-
-    /// Return the `original_filename` column for `id`, or `None` if no row exists.
-    ///
-    /// Used by `ImmichLibrary` to determine the file extension for cached originals.
-    pub async fn media_original_filename(
-        &self,
-        id: &MediaId,
-    ) -> Result<Option<String>, LibraryError> {
-        let id_str = id.as_str();
-        let row: Option<String> =
-            sqlx::query_scalar("SELECT original_filename FROM media WHERE id = ?")
-                .bind(id_str)
-                .fetch_optional(&self.pool)
-                .await
-                .map_err(LibraryError::Db)?;
-        Ok(row)
-    }
-
-    /// Used by `LocalLibrary` to construct the absolute original-file path.
-    pub async fn media_relative_path(
-        &self,
-        id: &MediaId,
-    ) -> Result<Option<String>, LibraryError> {
-        let id_str = id.as_str();
-        let row: Option<String> =
-            sqlx::query_scalar("SELECT relative_path FROM media WHERE id = ?")
-                .bind(id_str)
-                .fetch_optional(&self.pool)
-                .await
-                .map_err(LibraryError::Db)?;
-        Ok(row)
-    }
-
-    /// Return the stored [`ThumbnailStatus`] for `id`, or `None` if no row exists.
-    pub async fn thumbnail_status(
-        &self,
-        id: &MediaId,
-    ) -> Result<Option<ThumbnailStatus>, LibraryError> {
-        let id_str = id.as_str();
-        let row: Option<i64> =
-            sqlx::query_scalar("SELECT status FROM thumbnails WHERE media_id = ?")
-                .bind(id_str)
-                .fetch_optional(&self.pool)
-                .await
-                .map_err(LibraryError::Db)?;
-        Ok(row.map(ThumbnailStatus::from_i64))
-    }
-}
-
-#[async_trait::async_trait]
-impl LibraryMedia for Database {
-    async fn get_media_item(&self, id: &MediaId) -> Result<Option<MediaItem>, LibraryError> {
-        // Delegates to the public method on Database (defined above).
-        Database::get_media_item(self, id).await
-    }
-
-    async fn media_exists(&self, id: &MediaId) -> Result<bool, LibraryError> {
-        let id_str = id.as_str();
-        let count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM media WHERE id = ?")
-            .bind(id_str)
-            .fetch_one(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(count > 0)
-    }
-
-    async fn insert_media(&self, record: &MediaRecord) -> Result<(), LibraryError> {
-        sqlx::query(
-            "INSERT INTO media (id, relative_path, original_filename, file_size, imported_at,
-                                media_type, taken_at, width, height, orientation, duration_ms)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-        )
-        .bind(record.id.as_str())
-        .bind(&record.relative_path)
-        .bind(&record.original_filename)
-        .bind(record.file_size)
-        .bind(record.imported_at)
-        .bind(record.media_type as i64)
-        .bind(record.taken_at)
-        .bind(record.width)
-        .bind(record.height)
-        .bind(record.orientation as i64)
-        .bind(record.duration_ms.map(|v| v as i64))
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn list_media(
-        &self,
-        filter: MediaFilter,
-        cursor: Option<&MediaCursor>,
-        limit: u32,
-    ) -> Result<Vec<MediaItem>, LibraryError> {
-        // Each filter defines its own WHERE clause and sort expression.
-        // RecentImports sorts by imported_at; all others by taken_at.
-        let (filter_clause, sort_expr) = match &filter {
-            MediaFilter::All => (" AND is_trashed = 0", "COALESCE(taken_at, 0)"),
-            MediaFilter::Favorites => (
-                " AND is_trashed = 0 AND is_favorite = 1",
-                "COALESCE(taken_at, 0)",
-            ),
-            MediaFilter::Trashed => (" AND is_trashed = 1", "COALESCE(taken_at, 0)"),
-            MediaFilter::RecentImports { .. } => (
-                " AND is_trashed = 0 AND imported_at > ?",
-                "imported_at",
-            ),
-            MediaFilter::Album { .. } => (
-                " AND is_trashed = 0 AND id IN (SELECT media_id FROM album_media WHERE album_id = ?)",
-                "COALESCE(taken_at, 0)",
-            ),
-        };
-
-        // Extra bind parameter for filters that need one.
-        let extra_bind: Option<String> = match &filter {
-            MediaFilter::RecentImports { since } => Some(since.to_string()),
-            MediaFilter::Album { album_id } => Some(album_id.as_str().to_owned()),
-            _ => None,
-        };
-
-        let columns = "id, taken_at, imported_at, original_filename,
-                        width, height, orientation, media_type, is_favorite,
-                        is_trashed, trashed_at, duration_ms";
-
-        let rows = match cursor {
-            None => {
-                let sql = format!(
-                    "SELECT {columns}
-                     FROM media
-                     WHERE 1=1{filter_clause}
-                     ORDER BY {sort_expr} DESC, id DESC
-                     LIMIT ?"
-                );
-                let mut q = sqlx::query_as::<_, MediaRow>(&sql);
-                if let Some(ref val) = extra_bind {
-                    q = q.bind(val.as_str());
-                }
-                q.bind(limit as i64)
-                    .fetch_all(&self.pool)
-                    .await
-                    .map_err(LibraryError::Db)?
-            }
-            Some(cur) => {
-                let sql = format!(
-                    "SELECT {columns}
-                     FROM media
-                     WHERE ({sort_expr} < ?
-                        OR ({sort_expr} = ? AND id < ?)){filter_clause}
-                     ORDER BY {sort_expr} DESC, id DESC
-                     LIMIT ?"
-                );
-                let mut q = sqlx::query_as::<_, MediaRow>(&sql)
-                    .bind(cur.sort_key)
-                    .bind(cur.sort_key)
-                    .bind(cur.id.as_str());
-                if let Some(ref val) = extra_bind {
-                    q = q.bind(val.as_str());
-                }
-                q.bind(limit as i64)
-                    .fetch_all(&self.pool)
-                    .await
-                    .map_err(LibraryError::Db)?
-            }
-        };
-
-        Ok(rows.into_iter().map(MediaRow::into_item).collect())
-    }
-
-    async fn media_metadata(
-        &self,
-        id: &MediaId,
-    ) -> Result<Option<MediaMetadataRecord>, LibraryError> {
-        let row: Option<MetadataRow> = sqlx::query_as::<_, MetadataRow>(
-            "SELECT camera_make, camera_model, lens_model, aperture, shutter_str,
-                    iso, focal_length, gps_lat, gps_lon, gps_alt, color_space
-             FROM media_metadata WHERE media_id = ?",
-        )
-        .bind(id.as_str())
-        .fetch_optional(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-
-        Ok(row.map(|r| MediaMetadataRecord {
-            media_id: id.clone(),
-            camera_make: r.camera_make,
-            camera_model: r.camera_model,
-            lens_model: r.lens_model,
-            aperture: r.aperture,
-            shutter_str: r.shutter_str,
-            iso: r.iso.map(|v| v as u32),
-            focal_length: r.focal_length,
-            gps_lat: r.gps_lat,
-            gps_lon: r.gps_lon,
-            gps_alt: r.gps_alt,
-            color_space: r.color_space,
-        }))
-    }
-
-    async fn set_favorite(
-        &self,
-        ids: &[MediaId],
-        favorite: bool,
-    ) -> Result<(), LibraryError> {
-        if ids.is_empty() {
-            return Ok(());
-        }
-        let value: i64 = if favorite { 1 } else { 0 };
-        let placeholders = id_placeholders(ids.len());
-        let sql = format!("UPDATE media SET is_favorite = ? WHERE id IN ({placeholders})");
-        let mut query = sqlx::query(&sql);
-        query = query.bind(value);
-        for id in ids {
-            query = query.bind(id.as_str());
-        }
-        query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn trash(&self, ids: &[MediaId]) -> Result<(), LibraryError> {
-        if ids.is_empty() {
-            return Ok(());
-        }
-        let now = chrono::Utc::now().timestamp();
-        let placeholders = id_placeholders(ids.len());
-        let sql = format!(
-            "UPDATE media SET is_trashed = 1, trashed_at = ? WHERE id IN ({placeholders})"
-        );
-        let mut query = sqlx::query(&sql);
-        query = query.bind(now);
-        for id in ids {
-            query = query.bind(id.as_str());
-        }
-        query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn restore(&self, ids: &[MediaId]) -> Result<(), LibraryError> {
-        if ids.is_empty() {
-            return Ok(());
-        }
-        let placeholders = id_placeholders(ids.len());
-        let sql = format!(
-            "UPDATE media SET is_trashed = 0, trashed_at = NULL WHERE id IN ({placeholders})"
-        );
-        let mut query = sqlx::query(&sql);
-        for id in ids {
-            query = query.bind(id.as_str());
-        }
-        query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn delete_permanently(&self, ids: &[MediaId]) -> Result<(), LibraryError> {
-        if ids.is_empty() {
-            return Ok(());
-        }
-        let placeholders = id_placeholders(ids.len());
-        // Delete all FK references, then the media rows — one query per table.
-        for (table, col) in [
-            ("media_metadata", "media_id"),
-            ("thumbnails", "media_id"),
-            ("album_media", "media_id"),
-            ("media", "id"),
-        ] {
-            let sql = format!("DELETE FROM {table} WHERE {col} IN ({placeholders})");
-            let mut query = sqlx::query(&sql);
-            for id in ids {
-                query = query.bind(id.as_str());
-            }
-            query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-        }
-        Ok(())
-    }
-
-    async fn expired_trash(&self, max_age_secs: i64) -> Result<Vec<MediaId>, LibraryError> {
-        let cutoff = chrono::Utc::now().timestamp() - max_age_secs;
-        let rows: Vec<(String,)> = sqlx::query_as(
-            "SELECT id FROM media WHERE is_trashed = 1 AND trashed_at < ?",
-        )
-        .bind(cutoff)
-        .fetch_all(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(rows.into_iter().map(|(id,)| MediaId::new(id)).collect())
-    }
-
-    async fn insert_media_metadata(
-        &self,
-        record: &MediaMetadataRecord,
-    ) -> Result<(), LibraryError> {
-        if !record.has_data() {
-            return Ok(());
-        }
-        sqlx::query(
-            "INSERT INTO media_metadata
-                (media_id, camera_make, camera_model, lens_model, aperture, shutter_str,
-                 iso, focal_length, gps_lat, gps_lon, gps_alt, color_space)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-             ON CONFLICT(media_id) DO UPDATE SET
-                camera_make = excluded.camera_make,
-                camera_model = excluded.camera_model,
-                lens_model = excluded.lens_model,
-                aperture = excluded.aperture,
-                shutter_str = excluded.shutter_str,
-                iso = excluded.iso,
-                focal_length = excluded.focal_length,
-                gps_lat = excluded.gps_lat,
-                gps_lon = excluded.gps_lon,
-                gps_alt = excluded.gps_alt,
-                color_space = excluded.color_space",
-        )
-        .bind(record.media_id.as_str())
-        .bind(&record.camera_make)
-        .bind(&record.camera_model)
-        .bind(&record.lens_model)
-        .bind(record.aperture)
-        .bind(&record.shutter_str)
-        .bind(record.iso.map(|v| v as i64))
-        .bind(record.focal_length)
-        .bind(record.gps_lat)
-        .bind(record.gps_lon)
-        .bind(record.gps_alt)
-        .bind(&record.color_space)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-}
-
-/// Internal row type for album queries.
-#[derive(sqlx::FromRow)]
-struct AlbumRow {
-    id: String,
-    name: String,
-    created_at: i64,
-    updated_at: i64,
-    media_count: i64,
-    cover_media_id: Option<String>,
-}
-
-#[async_trait::async_trait]
-impl LibraryAlbums for Database {
-    async fn list_albums(&self) -> Result<Vec<Album>, LibraryError> {
-        let rows: Vec<AlbumRow> = sqlx::query_as(
-            "SELECT a.id, a.name, a.created_at, a.updated_at,
-                    COUNT(am.media_id) as media_count,
-                    (SELECT am2.media_id FROM album_media am2
-                     JOIN media m ON m.id = am2.media_id AND m.is_trashed = 0
-                     WHERE am2.album_id = a.id
-                     ORDER BY am2.added_at DESC LIMIT 1) as cover_media_id
-             FROM albums a
-             LEFT JOIN album_media am ON a.id = am.album_id
-                 LEFT JOIN media m2 ON am.media_id = m2.id AND m2.is_trashed = 0
-             GROUP BY a.id
-             ORDER BY a.updated_at DESC",
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-
-        Ok(rows
-            .into_iter()
-            .map(|r| Album {
-                id: AlbumId::from_raw(r.id),
-                name: r.name,
-                created_at: r.created_at,
-                updated_at: r.updated_at,
-                media_count: r.media_count as u32,
-                cover_media_id: r.cover_media_id.map(MediaId::new),
-            })
-            .collect())
-    }
-
-    async fn create_album(&self, name: &str) -> Result<AlbumId, LibraryError> {
-        let id = AlbumId::new();
-        let now = chrono::Utc::now().timestamp();
-        sqlx::query("INSERT INTO albums (id, name, created_at, updated_at) VALUES (?, ?, ?, ?)")
-            .bind(id.as_str())
-            .bind(name)
-            .bind(now)
-            .bind(now)
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(id)
-    }
-
-    async fn rename_album(&self, id: &AlbumId, name: &str) -> Result<(), LibraryError> {
-        let now = chrono::Utc::now().timestamp();
-        sqlx::query("UPDATE albums SET name = ?, updated_at = ? WHERE id = ?")
-            .bind(name)
-            .bind(now)
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn delete_album(&self, id: &AlbumId) -> Result<(), LibraryError> {
-        sqlx::query("DELETE FROM album_media WHERE album_id = ?")
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        sqlx::query("DELETE FROM albums WHERE id = ?")
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn add_to_album(
-        &self,
-        album_id: &AlbumId,
-        media_ids: &[MediaId],
-    ) -> Result<(), LibraryError> {
-        if media_ids.is_empty() {
-            return Ok(());
-        }
-        let now = chrono::Utc::now().timestamp();
-        // Batch insert via multi-row VALUES.
-        let row_placeholders: Vec<&str> = media_ids.iter().map(|_| "(?, ?, ?)").collect();
-        let sql = format!(
-            "INSERT OR IGNORE INTO album_media (album_id, media_id, added_at) VALUES {}",
-            row_placeholders.join(", ")
-        );
-        let mut query = sqlx::query(&sql);
-        for media_id in media_ids {
-            query = query
-                .bind(album_id.as_str())
-                .bind(media_id.as_str())
-                .bind(now);
-        }
-        query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-
-        sqlx::query("UPDATE albums SET updated_at = ? WHERE id = ?")
-            .bind(now)
-            .bind(album_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn remove_from_album(
-        &self,
-        album_id: &AlbumId,
-        media_ids: &[MediaId],
-    ) -> Result<(), LibraryError> {
-        if media_ids.is_empty() {
-            return Ok(());
-        }
-        let placeholders = id_placeholders(media_ids.len());
-        let sql = format!(
-            "DELETE FROM album_media WHERE album_id = ? AND media_id IN ({placeholders})"
-        );
-        let mut query = sqlx::query(&sql);
-        query = query.bind(album_id.as_str());
-        for media_id in media_ids {
-            query = query.bind(media_id.as_str());
-        }
-        query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-        let now = chrono::Utc::now().timestamp();
-        sqlx::query("UPDATE albums SET updated_at = ? WHERE id = ?")
-            .bind(now)
-            .bind(album_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    async fn list_album_media(
-        &self,
-        album_id: &AlbumId,
-        cursor: Option<&MediaCursor>,
-        limit: u32,
-    ) -> Result<Vec<MediaItem>, LibraryError> {
-        let rows = match cursor {
-            None => {
-                sqlx::query_as::<_, MediaRow>(
-                    "SELECT m.id, m.taken_at, m.imported_at, m.original_filename,
-                            m.width, m.height, m.orientation, m.media_type, m.is_favorite,
-                            m.is_trashed, m.trashed_at, m.duration_ms
-                     FROM media m
-                     JOIN album_media am ON m.id = am.media_id
-                     WHERE am.album_id = ? AND m.is_trashed = 0
-                     ORDER BY COALESCE(m.taken_at, 0) DESC, m.id DESC
-                     LIMIT ?",
-                )
-                .bind(album_id.as_str())
-                .bind(limit as i64)
-                .fetch_all(&self.pool)
-                .await
-                .map_err(LibraryError::Db)?
-            }
-            Some(cur) => {
-                sqlx::query_as::<_, MediaRow>(
-                    "SELECT m.id, m.taken_at, m.imported_at, m.original_filename,
-                            m.width, m.height, m.orientation, m.media_type, m.is_favorite,
-                            m.is_trashed, m.trashed_at, m.duration_ms
-                     FROM media m
-                     JOIN album_media am ON m.id = am.media_id
-                     WHERE am.album_id = ?
-                       AND (COALESCE(m.taken_at, 0) < ?
-                            OR (COALESCE(m.taken_at, 0) = ? AND m.id < ?))
-                       AND m.is_trashed = 0
-                     ORDER BY COALESCE(m.taken_at, 0) DESC, m.id DESC
-                     LIMIT ?",
-                )
-                .bind(album_id.as_str())
-                .bind(cur.sort_key)
-                .bind(cur.sort_key)
-                .bind(cur.id.as_str())
-                .bind(limit as i64)
-                .fetch_all(&self.pool)
-                .await
-                .map_err(LibraryError::Db)?
-            }
-        };
-
-        Ok(rows.into_iter().map(MediaRow::into_item).collect())
-    }
-}
-
-// ── Sync support (used by SyncManager for Immich backend) ───────────────────
-
-impl Database {
-    /// Upsert a media record. Uses `INSERT OR REPLACE` so existing records
-    /// are updated without a separate EXISTS check.
-    pub async fn upsert_media(&self, record: &MediaRecord) -> Result<(), LibraryError> {
-        sqlx::query(
-            "INSERT OR REPLACE INTO media (id, relative_path, original_filename, file_size,
-                                           imported_at, media_type, taken_at, width, height,
-                                           orientation, duration_ms, is_favorite, is_trashed,
-                                           trashed_at)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-        )
-        .bind(record.id.as_str())
-        .bind(&record.relative_path)
-        .bind(&record.original_filename)
-        .bind(record.file_size)
-        .bind(record.imported_at)
-        .bind(record.media_type as i64)
-        .bind(record.taken_at)
-        .bind(record.width)
-        .bind(record.height)
-        .bind(record.orientation as i64)
-        .bind(record.duration_ms.map(|v| v as i64))
-        .bind(record.is_favorite as i64)
-        .bind(record.is_trashed as i64)
-        .bind(record.trashed_at)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Upsert a media metadata record.
-    pub async fn upsert_media_metadata(
-        &self,
-        record: &MediaMetadataRecord,
-    ) -> Result<(), LibraryError> {
-        if !record.has_data() {
-            return Ok(());
-        }
-        sqlx::query(
-            "INSERT OR REPLACE INTO media_metadata
-                (media_id, camera_make, camera_model, lens_model, aperture, shutter_str,
-                 iso, focal_length, gps_lat, gps_lon, gps_alt, color_space)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-        )
-        .bind(record.media_id.as_str())
-        .bind(&record.camera_make)
-        .bind(&record.camera_model)
-        .bind(&record.lens_model)
-        .bind(record.aperture)
-        .bind(&record.shutter_str)
-        .bind(record.iso.map(|v| v as i64))
-        .bind(record.focal_length)
-        .bind(record.gps_lat)
-        .bind(record.gps_lon)
-        .bind(record.gps_alt)
-        .bind(&record.color_space)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Upsert an album record from the sync stream.
-    pub async fn upsert_album(
-        &self,
-        id: &str,
-        name: &str,
-        created_at: i64,
-        updated_at: i64,
-    ) -> Result<(), LibraryError> {
-        sqlx::query(
-            "INSERT OR REPLACE INTO albums (id, name, created_at, updated_at) VALUES (?, ?, ?, ?)",
-        )
-        .bind(id)
-        .bind(name)
-        .bind(created_at)
-        .bind(updated_at)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Upsert an album-media association from the sync stream.
-    pub async fn upsert_album_media(
-        &self,
-        album_id: &str,
-        media_id: &str,
-        added_at: i64,
-    ) -> Result<(), LibraryError> {
-        sqlx::query(
-            "INSERT OR IGNORE INTO album_media (album_id, media_id, added_at) VALUES (?, ?, ?)",
-        )
-        .bind(album_id)
-        .bind(media_id)
-        .bind(added_at)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Delete a single album-media association.
-    pub async fn delete_album_media_entry(
-        &self,
-        album_id: &str,
-        media_id: &str,
-    ) -> Result<(), LibraryError> {
-        sqlx::query("DELETE FROM album_media WHERE album_id = ? AND media_id = ?")
-            .bind(album_id)
-            .bind(media_id)
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Load all media IDs into a HashSet (for reset sync deletion detection).
-    pub async fn all_media_ids(&self) -> Result<std::collections::HashSet<String>, LibraryError> {
-        let rows: Vec<(String,)> = sqlx::query_as("SELECT id FROM media")
-            .fetch_all(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(rows.into_iter().map(|(id,)| id).collect())
-    }
-
-    /// Retrieve all sync checkpoints.
-    pub async fn get_sync_checkpoints(
-        &self,
-    ) -> Result<std::collections::HashMap<String, String>, LibraryError> {
-        let rows: Vec<(String, String)> =
-            sqlx::query_as("SELECT entity_type, ack FROM sync_checkpoints")
-                .fetch_all(&self.pool)
-                .await
-                .map_err(LibraryError::Db)?;
-        Ok(rows.into_iter().collect())
-    }
-
-    /// Save sync checkpoints (batch upsert).
-    pub async fn save_sync_checkpoints(
-        &self,
-        acks: &[(String, String)],
-    ) -> Result<(), LibraryError> {
-        if acks.is_empty() {
-            return Ok(());
-        }
-        let row_placeholders: Vec<&str> = acks.iter().map(|_| "(?, ?)").collect();
-        let sql = format!(
-            "INSERT OR REPLACE INTO sync_checkpoints (entity_type, ack) VALUES {}",
-            row_placeholders.join(", ")
-        );
-        let mut query = sqlx::query(&sql);
-        for (entity_type, ack) in acks {
-            query = query.bind(entity_type).bind(ack);
-        }
-        query.execute(&self.pool).await.map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Clear all sync checkpoints (for reset sync).
-    pub async fn clear_sync_checkpoints(&self) -> Result<(), LibraryError> {
-        sqlx::query("DELETE FROM sync_checkpoints")
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    // ── Library stats ─────────────────────────────────────────────────────
-
-    /// Return aggregate library statistics for the preferences overview.
-    pub async fn library_stats(&self) -> Result<LibraryStats, LibraryError> {
-        let row: (i64, i64, i64) = sqlx::query_as(
-            "SELECT
-                COUNT(CASE WHEN media_type = 0 AND is_trashed = 0 THEN 1 END),
-                COUNT(CASE WHEN media_type = 1 AND is_trashed = 0 THEN 1 END),
-                COALESCE(SUM(CASE WHEN is_trashed = 0 THEN file_size ELSE 0 END), 0)
-             FROM media",
-        )
-        .fetch_one(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-
-        let album_count: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM albums")
-            .fetch_one(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-
-        Ok(LibraryStats {
-            photo_count: row.0 as u64,
-            video_count: row.1 as u64,
-            album_count: album_count.0 as u64,
-            total_file_size: row.2 as u64,
-        })
-    }
-
-    // ── Upload queue (Immich import crash recovery) ─────────────────────────
-
-    /// Insert a file into the upload queue as pending.
-    pub async fn insert_upload_pending(
-        &self,
-        file_path: &str,
-        created_at: i64,
-    ) -> Result<(), LibraryError> {
-        sqlx::query(
-            "INSERT OR IGNORE INTO upload_queue (file_path, status, created_at) VALUES (?, 0, ?)",
-        )
-        .bind(file_path)
-        .bind(created_at)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Store the computed SHA-1 hash for a queued file.
-    pub async fn set_upload_hash(
-        &self,
-        file_path: &str,
-        sha1_hash: &str,
-    ) -> Result<(), LibraryError> {
-        sqlx::query("UPDATE upload_queue SET sha1_hash = ? WHERE file_path = ?")
-            .bind(sha1_hash)
-            .bind(file_path)
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Update the status of a queued upload.
-    pub async fn set_upload_status(
-        &self,
-        file_path: &str,
-        status: i64,
-        error_msg: Option<&str>,
-    ) -> Result<(), LibraryError> {
-        sqlx::query("UPDATE upload_queue SET status = ?, error_msg = ? WHERE file_path = ?")
-            .bind(status)
-            .bind(error_msg)
-            .bind(file_path)
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Return all pending uploads (status = 0) for resume after crash.
-    pub async fn get_pending_uploads(
-        &self,
-    ) -> Result<Vec<(String, Option<String>)>, LibraryError> {
-        let rows: Vec<(String, Option<String>)> = sqlx::query_as(
-            "SELECT file_path, sha1_hash FROM upload_queue WHERE status = 0 ORDER BY created_at",
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(rows)
-    }
-
-    /// Remove all completed/duplicate uploads from the queue.
-    pub async fn clear_completed_uploads(&self) -> Result<(), LibraryError> {
-        sqlx::query("DELETE FROM upload_queue WHERE status IN (1, 3)")
-            .execute(&self.pool)
-            .await
-            .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    // ── Sync audit ──────────────────────────────────────────────────────────
-
-    /// Record the start of processing a sync record.
-    /// Returns the row id for later completion via [`complete_sync_audit`].
-    pub async fn start_sync_audit(
-        &self,
-        entity_type: &str,
-        entity_id: &str,
-        sync_cycle: &str,
-    ) -> Result<i64, LibraryError> {
-        let now = chrono::Utc::now().to_rfc3339();
-        let result = sqlx::query(
-            "INSERT INTO sync_audit (entity_type, entity_id, action, started_at, sync_cycle)
-             VALUES (?, ?, 'upsert', ?, ?)",
-        )
-        .bind(entity_type)
-        .bind(entity_id)
-        .bind(&now)
-        .bind(sync_cycle)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(result.last_insert_rowid())
-    }
-
-    /// Mark a sync audit record as completed (just before acking).
-    pub async fn complete_sync_audit(
-        &self,
-        row_id: i64,
-        action: &str,
-    ) -> Result<(), LibraryError> {
-        let now = chrono::Utc::now().to_rfc3339();
-        sqlx::query(
-            "UPDATE sync_audit SET completed_at = ?, action = ? WHERE id = ?",
-        )
-        .bind(&now)
-        .bind(action)
-        .bind(row_id)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-
-    /// Mark a sync audit record as failed with an error message.
-    pub async fn fail_sync_audit(
-        &self,
-        row_id: i64,
-        error_msg: &str,
-    ) -> Result<(), LibraryError> {
-        let now = chrono::Utc::now().to_rfc3339();
-        sqlx::query(
-            "UPDATE sync_audit SET completed_at = ?, action = 'error', error_msg = ? WHERE id = ?",
-        )
-        .bind(&now)
-        .bind(error_msg)
-        .bind(row_id)
-        .execute(&self.pool)
-        .await
-        .map_err(LibraryError::Db)?;
-        Ok(())
-    }
-}
-
 /// Build a comma-separated list of `?` placeholders for an `IN (...)` clause.
 fn id_placeholders(count: usize) -> String {
     let mut s = String::with_capacity(count * 3);
@@ -1068,17 +85,17 @@ fn id_placeholders(count: usize) -> String {
     s
 }
 
+/// Shared test utilities for all db submodules.
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::library::media::{LibraryMedia, MediaType};
-    use tempfile::tempdir;
+pub(crate) mod test_helpers {
+    use super::Database;
+    use crate::library::media::{MediaId, MediaRecord, MediaType};
 
-    async fn open_test_db(dir: &std::path::Path) -> Database {
+    pub async fn open_test_db(dir: &std::path::Path) -> Database {
         Database::open(&dir.join("test.db")).await.unwrap()
     }
 
-    fn test_record(id: MediaId) -> MediaRecord {
+    pub fn test_record(id: MediaId) -> MediaRecord {
         MediaRecord {
             id,
             relative_path: "2025/01/15/photo.jpg".to_string(),
@@ -1097,48 +114,7 @@ mod tests {
         }
     }
 
-    #[tokio::test]
-    async fn open_creates_database_file() {
-        let dir = tempdir().unwrap();
-        let db_path = dir.path().join("sub").join("moments.db");
-        Database::open(&db_path).await.unwrap();
-        assert!(db_path.exists());
-    }
-
-    #[tokio::test]
-    async fn media_does_not_exist_initially() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::from_file(std::path::Path::new(file!())).await.unwrap();
-        assert!(!db.media_exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn insert_and_exists_roundtrip() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::from_file(std::path::Path::new(file!())).await.unwrap();
-
-        db.insert_media(&test_record(id.clone())).await.unwrap();
-
-        assert!(db.media_exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn duplicate_insert_returns_error() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::from_file(std::path::Path::new(file!())).await.unwrap();
-
-        let record = test_record(id.clone());
-
-        db.insert_media(&record).await.unwrap();
-        assert!(db.insert_media(&record).await.is_err());
-    }
-
-    // ── list_media tests ──────────────────────────────────────────────────────
-
-    fn record_with_taken_at(id: MediaId, path: &str, taken_at: Option<i64>) -> MediaRecord {
+    pub fn record_with_taken_at(id: MediaId, path: &str, taken_at: Option<i64>) -> MediaRecord {
         MediaRecord {
             id,
             relative_path: path.to_string(),
@@ -1157,271 +133,7 @@ mod tests {
         }
     }
 
-    #[tokio::test]
-    async fn list_media_empty_returns_empty_vec() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let result = db.list_media(MediaFilter::All, None, 50).await.unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[tokio::test]
-    async fn list_media_first_page_ordered_reverse_chronological() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        // Insert three items with different taken_at values.
-        let id_a = MediaId::new("a".repeat(64));
-        let id_b = MediaId::new("b".repeat(64));
-        let id_c = MediaId::new("c".repeat(64));
-
-        db.insert_media(&record_with_taken_at(id_a.clone(), "2025/01/01/a.jpg", Some(1_000))).await.unwrap();
-        db.insert_media(&record_with_taken_at(id_b.clone(), "2025/01/02/b.jpg", Some(3_000))).await.unwrap();
-        db.insert_media(&record_with_taken_at(id_c.clone(), "2025/01/03/c.jpg", Some(2_000))).await.unwrap();
-
-        let items = db.list_media(MediaFilter::All, None, 50).await.unwrap();
-
-        assert_eq!(items.len(), 3);
-        // Newest first: b (3000) → c (2000) → a (1000)
-        assert_eq!(items[0].id, id_b);
-        assert_eq!(items[1].id, id_c);
-        assert_eq!(items[2].id, id_a);
-    }
-
-    #[tokio::test]
-    async fn list_media_null_taken_at_sorts_to_end() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let id_dated = MediaId::new("d".repeat(64));
-        let id_undated = MediaId::new("e".repeat(64));
-
-        db.insert_media(&record_with_taken_at(id_dated.clone(), "2025/01/01/dated.jpg", Some(5_000))).await.unwrap();
-        db.insert_media(&record_with_taken_at(id_undated.clone(), "2025/01/01/undated.jpg", None)).await.unwrap();
-
-        let items = db.list_media(MediaFilter::All, None, 50).await.unwrap();
-        assert_eq!(items.len(), 2);
-        // Dated item first, undated last.
-        assert_eq!(items[0].id, id_dated);
-        assert_eq!(items[1].id, id_undated);
-    }
-
-    #[tokio::test]
-    async fn list_media_cursor_returns_next_page() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        // Insert 5 items with descending timestamps.
-        let ids: Vec<MediaId> = (1..=5)
-            .map(|i| MediaId::new(format!("{:0>64}", i)))
-            .collect();
-
-        for (i, id) in ids.iter().enumerate() {
-            let ts = (5 - i as i64) * 1000; // 5000, 4000, 3000, 2000, 1000
-            db.insert_media(&record_with_taken_at(
-                id.clone(),
-                &format!("2025/01/0{}/photo.jpg", i + 1),
-                Some(ts),
-            ))
-            .await
-            .unwrap();
-        }
-
-        // First page: 3 items (newest 3: ids[0]=5000, ids[1]=4000, ids[2]=3000)
-        let page1 = db.list_media(MediaFilter::All, None, 3).await.unwrap();
-        assert_eq!(page1.len(), 3);
-        assert_eq!(page1[0].taken_at, Some(5000));
-        assert_eq!(page1[2].taken_at, Some(3000));
-
-        // Build cursor from last item of page 1.
-        let last = &page1[2];
-        let cursor = MediaCursor {
-            sort_key: last.taken_at.unwrap_or(0),
-            id: last.id.clone(),
-        };
-
-        // Second page: remaining 2 items (ids[3]=2000, ids[4]=1000)
-        let page2 = db.list_media(MediaFilter::All, Some(&cursor), 3).await.unwrap();
-        assert_eq!(page2.len(), 2);
-        assert_eq!(page2[0].taken_at, Some(2000));
-        assert_eq!(page2[1].taken_at, Some(1000));
-    }
-
-    #[tokio::test]
-    async fn list_media_respects_limit() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        for i in 0..10u64 {
-            let id = MediaId::new(format!("{i:0>64}"));
-            db.insert_media(&record_with_taken_at(
-                id,
-                &format!("2025/01/{i:02}/photo.jpg"),
-                Some(i as i64 * 1000),
-            ))
-            .await
-            .unwrap();
-        }
-
-        let items = db.list_media(MediaFilter::All, None, 4).await.unwrap();
-        assert_eq!(items.len(), 4);
-    }
-
-    // ── set_favorite tests ───────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn set_favorite_and_read_back() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::new("f".repeat(64));
-        db.insert_media(&test_record(id.clone())).await.unwrap();
-
-        // Initially not a favourite.
-        let items = db.list_media(MediaFilter::All, None, 10).await.unwrap();
-        assert!(!items[0].is_favorite);
-
-        // Set favourite.
-        db.set_favorite(&[id.clone()], true).await.unwrap();
-        let items = db.list_media(MediaFilter::All, None, 10).await.unwrap();
-        assert!(items[0].is_favorite);
-
-        // Clear favourite.
-        db.set_favorite(&[id], false).await.unwrap();
-        let items = db.list_media(MediaFilter::All, None, 10).await.unwrap();
-        assert!(!items[0].is_favorite);
-    }
-
-    #[tokio::test]
-    async fn set_favorite_multiple_ids() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let id1 = MediaId::new("1".repeat(64));
-        let id2 = MediaId::new("2".repeat(64));
-        db.insert_media(&record_with_taken_at(id1.clone(), "a.jpg", Some(1000)))
-            .await
-            .unwrap();
-        db.insert_media(&record_with_taken_at(id2.clone(), "b.jpg", Some(2000)))
-            .await
-            .unwrap();
-
-        db.set_favorite(&[id1.clone(), id2.clone()], true)
-            .await
-            .unwrap();
-
-        let items = db.list_media(MediaFilter::All, None, 10).await.unwrap();
-        assert!(items.iter().all(|i| i.is_favorite));
-    }
-
-    #[tokio::test]
-    async fn list_media_favorites_filter() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let id1 = MediaId::new("1".repeat(64));
-        let id2 = MediaId::new("2".repeat(64));
-        db.insert_media(&record_with_taken_at(id1.clone(), "a.jpg", Some(1000)))
-            .await
-            .unwrap();
-        db.insert_media(&record_with_taken_at(id2.clone(), "b.jpg", Some(2000)))
-            .await
-            .unwrap();
-
-        // Mark only id1 as favourite.
-        db.set_favorite(&[id1.clone()], true).await.unwrap();
-
-        // All filter returns both.
-        let all = db.list_media(MediaFilter::All, None, 10).await.unwrap();
-        assert_eq!(all.len(), 2);
-
-        // Favorites filter returns only id1.
-        let favs = db.list_media(MediaFilter::Favorites, None, 10).await.unwrap();
-        assert_eq!(favs.len(), 1);
-        assert_eq!(favs[0].id, id1);
-    }
-
-    // ── trash tests ──────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn trash_and_restore_roundtrip() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::new("t".repeat(64));
-        db.insert_media(&test_record(id.clone())).await.unwrap();
-
-        // Initially visible in All.
-        assert_eq!(db.list_media(MediaFilter::All, None, 10).await.unwrap().len(), 1);
-
-        // Trash it.
-        db.trash(&[id.clone()]).await.unwrap();
-        assert_eq!(db.list_media(MediaFilter::All, None, 10).await.unwrap().len(), 0);
-        let trashed = db.list_media(MediaFilter::Trashed, None, 10).await.unwrap();
-        assert_eq!(trashed.len(), 1);
-        assert!(trashed[0].is_trashed);
-        assert!(trashed[0].trashed_at.is_some());
-
-        // Restore it.
-        db.restore(&[id]).await.unwrap();
-        assert_eq!(db.list_media(MediaFilter::All, None, 10).await.unwrap().len(), 1);
-        assert_eq!(db.list_media(MediaFilter::Trashed, None, 10).await.unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn trash_excludes_from_favorites() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::new("u".repeat(64));
-        db.insert_media(&test_record(id.clone())).await.unwrap();
-        db.set_favorite(&[id.clone()], true).await.unwrap();
-
-        assert_eq!(db.list_media(MediaFilter::Favorites, None, 10).await.unwrap().len(), 1);
-
-        db.trash(&[id]).await.unwrap();
-        assert_eq!(db.list_media(MediaFilter::Favorites, None, 10).await.unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn delete_permanently_removes_row() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::new("v".repeat(64));
-        db.insert_media(&test_record(id.clone())).await.unwrap();
-
-        db.delete_permanently(&[id.clone()]).await.unwrap();
-        assert!(!db.media_exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn expired_trash_returns_old_items() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-        let id = MediaId::new("w".repeat(64));
-        db.insert_media(&test_record(id.clone())).await.unwrap();
-
-        db.trash(&[id.clone()]).await.unwrap();
-
-        // Item was just trashed — not expired with any positive max_age.
-        let expired = db.expired_trash(30 * 24 * 60 * 60).await.unwrap();
-        assert!(expired.is_empty());
-
-        // Manually backdate trashed_at to 31 days ago.
-        let old_ts = chrono::Utc::now().timestamp() - (31 * 24 * 60 * 60);
-        sqlx::query("UPDATE media SET trashed_at = ? WHERE id = ?")
-            .bind(old_ts)
-            .bind(id.as_str())
-            .execute(&db.pool)
-            .await
-            .unwrap();
-
-        // Now it's expired with a 30-day window.
-        let expired = db.expired_trash(30 * 24 * 60 * 60).await.unwrap();
-        assert_eq!(expired.len(), 1);
-        assert_eq!(expired[0], id);
-    }
-
-    // ── recent imports tests ─────────────────────────────────────────────────
-
-    fn record_with_imported_at(id: MediaId, path: &str, imported_at: i64) -> MediaRecord {
+    pub fn record_with_imported_at(id: MediaId, path: &str, imported_at: i64) -> MediaRecord {
         MediaRecord {
             id,
             relative_path: path.to_string(),
@@ -1439,348 +151,18 @@ mod tests {
             trashed_at: None,
         }
     }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tempfile::tempdir;
 
     #[tokio::test]
-    async fn list_media_recent_imports_filter() {
+    async fn open_creates_database_file() {
         let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let now = chrono::Utc::now().timestamp();
-        let recent = now - 3600; // 1 hour ago
-        let old = now - 90 * 86400; // 90 days ago
-
-        let id_recent = MediaId::new("r".repeat(64));
-        let id_old = MediaId::new("o".repeat(64));
-
-        db.insert_media(&record_with_imported_at(id_recent.clone(), "recent.jpg", recent))
-            .await
-            .unwrap();
-        db.insert_media(&record_with_imported_at(id_old.clone(), "old.jpg", old))
-            .await
-            .unwrap();
-
-        let since = now - 30 * 86400; // last 30 days
-        let items = db
-            .list_media(MediaFilter::RecentImports { since }, None, 50)
-            .await
-            .unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].id, id_recent);
-    }
-
-    #[tokio::test]
-    async fn list_media_recent_imports_excludes_trashed() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let now = chrono::Utc::now().timestamp();
-        let id = MediaId::new("x".repeat(64));
-        db.insert_media(&record_with_imported_at(id.clone(), "trashed.jpg", now - 3600))
-            .await
-            .unwrap();
-        db.trash(&[id]).await.unwrap();
-
-        let since = now - 30 * 86400;
-        let items = db
-            .list_media(MediaFilter::RecentImports { since }, None, 50)
-            .await
-            .unwrap();
-
-        assert!(items.is_empty());
-    }
-
-    #[tokio::test]
-    async fn list_media_recent_imports_sorted_by_imported_at() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let now = chrono::Utc::now().timestamp();
-        let id_first = MediaId::new("1".repeat(64));
-        let id_second = MediaId::new("2".repeat(64));
-
-        // id_first imported earlier, id_second imported later
-        db.insert_media(&record_with_imported_at(id_first.clone(), "first.jpg", now - 7200))
-            .await
-            .unwrap();
-        db.insert_media(&record_with_imported_at(id_second.clone(), "second.jpg", now - 3600))
-            .await
-            .unwrap();
-
-        let since = now - 30 * 86400;
-        let items = db
-            .list_media(MediaFilter::RecentImports { since }, None, 50)
-            .await
-            .unwrap();
-
-        assert_eq!(items.len(), 2);
-        // Most recently imported first
-        assert_eq!(items[0].id, id_second);
-        assert_eq!(items[1].id, id_first);
-    }
-
-    // ── album tests ──────────────────────────────────────────────────────────
-
-    use crate::library::album::LibraryAlbums;
-
-    #[tokio::test]
-    async fn create_album_and_list() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let id = db.create_album("Vacation").await.unwrap();
-        let albums = db.list_albums().await.unwrap();
-
-        assert_eq!(albums.len(), 1);
-        assert_eq!(albums[0].id, id);
-        assert_eq!(albums[0].name, "Vacation");
-        assert_eq!(albums[0].media_count, 0);
-        assert!(albums[0].cover_media_id.is_none());
-    }
-
-    #[tokio::test]
-    async fn create_album_generates_unique_ids() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let id1 = db.create_album("Album 1").await.unwrap();
-        let id2 = db.create_album("Album 2").await.unwrap();
-        assert_ne!(id1, id2);
-    }
-
-    #[tokio::test]
-    async fn rename_album_updates_name() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let id = db.create_album("Old Name").await.unwrap();
-        db.rename_album(&id, "New Name").await.unwrap();
-
-        let albums = db.list_albums().await.unwrap();
-        assert_eq!(albums[0].name, "New Name");
-    }
-
-    #[tokio::test]
-    async fn delete_album_removes_album_and_media_links() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("To Delete").await.unwrap();
-        let media_id = MediaId::new("d".repeat(64));
-        db.insert_media(&test_record(media_id.clone())).await.unwrap();
-        db.add_to_album(&album_id, &[media_id.clone()]).await.unwrap();
-
-        db.delete_album(&album_id).await.unwrap();
-
-        let albums = db.list_albums().await.unwrap();
-        assert!(albums.is_empty());
-        // Media item itself still exists.
-        assert!(db.media_exists(&media_id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn add_to_album_and_list_media() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("My Album").await.unwrap();
-        let id_a = MediaId::new("a".repeat(64));
-        let id_b = MediaId::new("b".repeat(64));
-        db.insert_media(&record_with_taken_at(id_a.clone(), "a.jpg", Some(1000)))
-            .await.unwrap();
-        db.insert_media(&record_with_taken_at(id_b.clone(), "b.jpg", Some(2000)))
-            .await.unwrap();
-
-        db.add_to_album(&album_id, &[id_a.clone(), id_b.clone()])
-            .await.unwrap();
-
-        let items = db.list_album_media(&album_id, None, 50).await.unwrap();
-        assert_eq!(items.len(), 2);
-        // Sorted by taken_at DESC: b (2000) first, a (1000) second.
-        assert_eq!(items[0].id, id_b);
-        assert_eq!(items[1].id, id_a);
-    }
-
-    #[tokio::test]
-    async fn add_duplicate_to_album_is_idempotent() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Dupes").await.unwrap();
-        let media_id = MediaId::new("e".repeat(64));
-        db.insert_media(&test_record(media_id.clone())).await.unwrap();
-
-        db.add_to_album(&album_id, &[media_id.clone()]).await.unwrap();
-        db.add_to_album(&album_id, &[media_id.clone()]).await.unwrap();
-
-        let items = db.list_album_media(&album_id, None, 50).await.unwrap();
-        assert_eq!(items.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn remove_from_album() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Remove Test").await.unwrap();
-        let id_a = MediaId::new("a".repeat(64));
-        let id_b = MediaId::new("b".repeat(64));
-        db.insert_media(&record_with_taken_at(id_a.clone(), "a.jpg", Some(1000)))
-            .await.unwrap();
-        db.insert_media(&record_with_taken_at(id_b.clone(), "b.jpg", Some(2000)))
-            .await.unwrap();
-        db.add_to_album(&album_id, &[id_a.clone(), id_b.clone()])
-            .await.unwrap();
-
-        db.remove_from_album(&album_id, &[id_a]).await.unwrap();
-
-        let items = db.list_album_media(&album_id, None, 50).await.unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].id, id_b);
-    }
-
-    #[tokio::test]
-    async fn list_albums_includes_media_count() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Counting").await.unwrap();
-        let id_a = MediaId::new("a".repeat(64));
-        let id_b = MediaId::new("b".repeat(64));
-        db.insert_media(&record_with_taken_at(id_a.clone(), "a.jpg", Some(1000)))
-            .await.unwrap();
-        db.insert_media(&record_with_taken_at(id_b.clone(), "b.jpg", Some(2000)))
-            .await.unwrap();
-        db.add_to_album(&album_id, &[id_a, id_b]).await.unwrap();
-
-        let albums = db.list_albums().await.unwrap();
-        assert_eq!(albums[0].media_count, 2);
-    }
-
-    #[tokio::test]
-    async fn list_albums_includes_cover_media_id() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Cover").await.unwrap();
-        let id_a = MediaId::new("a".repeat(64));
-        db.insert_media(&record_with_taken_at(id_a.clone(), "a.jpg", Some(1000)))
-            .await.unwrap();
-        db.add_to_album(&album_id, &[id_a.clone()]).await.unwrap();
-
-        let albums = db.list_albums().await.unwrap();
-        assert_eq!(albums[0].cover_media_id, Some(id_a));
-    }
-
-    #[tokio::test]
-    async fn list_album_media_excludes_trashed() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Trash Test").await.unwrap();
-        let media_id = MediaId::new("t".repeat(64));
-        db.insert_media(&test_record(media_id.clone())).await.unwrap();
-        db.add_to_album(&album_id, &[media_id.clone()]).await.unwrap();
-
-        db.trash(&[media_id]).await.unwrap();
-
-        let items = db.list_album_media(&album_id, None, 50).await.unwrap();
-        assert!(items.is_empty());
-    }
-
-    #[tokio::test]
-    async fn list_album_media_cursor_pagination() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Paging").await.unwrap();
-        let ids: Vec<MediaId> = (1..=5)
-            .map(|i| MediaId::new(format!("{:0>64}", i)))
-            .collect();
-
-        for (i, id) in ids.iter().enumerate() {
-            let ts = (5 - i as i64) * 1000;
-            db.insert_media(&record_with_taken_at(
-                id.clone(),
-                &format!("2025/01/0{}/photo.jpg", i + 1),
-                Some(ts),
-            ))
-            .await
-            .unwrap();
-        }
-        let all_ids: Vec<MediaId> = ids.clone();
-        db.add_to_album(&album_id, &all_ids).await.unwrap();
-
-        // First page: 3 items.
-        let page1 = db.list_album_media(&album_id, None, 3).await.unwrap();
-        assert_eq!(page1.len(), 3);
-
-        let last = &page1[2];
-        let cursor = MediaCursor {
-            sort_key: last.taken_at.unwrap_or(0),
-            id: last.id.clone(),
-        };
-
-        // Second page: remaining 2 items.
-        let page2 = db.list_album_media(&album_id, Some(&cursor), 3).await.unwrap();
-        assert_eq!(page2.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn delete_media_removes_from_album() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Cascade").await.unwrap();
-        let media_id = MediaId::new("c".repeat(64));
-        db.insert_media(&test_record(media_id.clone())).await.unwrap();
-        db.add_to_album(&album_id, &[media_id.clone()]).await.unwrap();
-
-        db.delete_permanently(&[media_id]).await.unwrap();
-
-        let items = db.list_album_media(&album_id, None, 50).await.unwrap();
-        assert!(items.is_empty());
-    }
-
-    // ── album via MediaFilter tests ──────────────────────────────────────
-
-    #[tokio::test]
-    async fn list_media_album_filter() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Filter Test").await.unwrap();
-        let id_in = MediaId::new("i".repeat(64));
-        let id_out = MediaId::new("o".repeat(64));
-        db.insert_media(&record_with_taken_at(id_in.clone(), "in.jpg", Some(2000)))
-            .await.unwrap();
-        db.insert_media(&record_with_taken_at(id_out.clone(), "out.jpg", Some(1000)))
-            .await.unwrap();
-        db.add_to_album(&album_id, &[id_in.clone()]).await.unwrap();
-
-        let items = db
-            .list_media(MediaFilter::Album { album_id: album_id.clone() }, None, 50)
-            .await
-            .unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].id, id_in);
-    }
-
-    #[tokio::test]
-    async fn list_media_album_filter_excludes_trashed() {
-        let dir = tempdir().unwrap();
-        let db = open_test_db(dir.path()).await;
-
-        let album_id = db.create_album("Trash Filter").await.unwrap();
-        let media_id = MediaId::new("t".repeat(64));
-        db.insert_media(&test_record(media_id.clone())).await.unwrap();
-        db.add_to_album(&album_id, &[media_id.clone()]).await.unwrap();
-        db.trash(&[media_id]).await.unwrap();
-
-        let items = db
-            .list_media(MediaFilter::Album { album_id }, None, 50)
-            .await
-            .unwrap();
-        assert!(items.is_empty());
+        let db_path = dir.path().join("sub").join("moments.db");
+        Database::open(&db_path).await.unwrap();
+        assert!(db_path.exists());
     }
 }
